refactor to exporting the interfaces of attach to process
